backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ravindu.somawansa <ravindu.somawansa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.12"
fastapi = "0.114.2"
mangum = "0.17.0"
python-jose = "^3.3.0"
requests = "^2.32.3"
boto3 = "^1.35.20"
pydantic = "^2.9.1"
fastapi-jwt-auth = "^0.5.0"
uvicorn = "^0.30.6"
bcrypt = "^4.2.0"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.35"
passlib = "^1.7.4"
email-validator = "^2.2.0"
apache-age-python = "^0.0.7"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.12"
langchain-openai = "^0.2.5"
langchain = "^0.3.7"
pillow = "^11.0.0"
pdf2image = "^1.17.0"
networkx = "^3.4.2" # dev
matplotlib = "^3.9.2" # dev
langchain-aws = "^0.2.7"
pymupdf = "^1.24.13"
langchain-anthropic = "^0.3.0"
asyncio = "^3.4.3"
watchfiles = "^1.0.3"
langgraph = "^0.2.60"
streamlit = "^1.41.1"
streamlit-mermaid = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
